Reorganize code into more sensible structure
