+++ src/config.rs
@@ -1,5 +1,6 @@
 use crate::error::Error;
-use crate::store::delay::RetryDelay;
+
+use crate::RetryDelay;
 
 const MAX_RETRIES_LIMIT: u32 = 10;
 const MIN_DELAY_MS: u64 = 50;
