+++ src/kurrent_adapter/stream.rs
